--- elements/pl-code-output/info.json
@@ -1,8 +1,0 @@
-{
-    "controller": "pl-code-output.py",
-    "dependencies": {
-        "elementStyles": [
-            "pl-code-output.css"
-        ]
-    }
-}
