+++ elements/pl-python-variable/info.json
@@ -1,0 +1,8 @@
+{
+    "controller": "pl-python-variable.py",
+    "dependencies": {
+        "elementStyles": [
+            "pl-python-variable.css"
+        ]
+    }
+}
