+++ exampleCourse/questions/examplesCodeOutput/info.json
@@ -1,6 +1,6 @@
 {
     "uuid": "815ff014-338c-4f0c-8672-56dcd0675969",
-    "title": "Rendering Python variables using pl-code-output",
+    "title": "Rendering Python variables using pl-python-variable",
     "topic": "Test",
     "type": "v3",
     "tags": [
